Moved the process code into a module
This MR moves the main code of updatesnap into a module, thus
allowing to reuse it in other tools.
